requirement.txt
pyttsx3==2.90
requests==2.27.1
websocket-client==1.3.1
urllib3==1.26.4